Split out and rename classes to make sense
