+++ main-module/src/main/java/com/diffblue/javatest/Search.java
@@ -1,7 +1,7 @@
 package com.diffblue.javatest;
 
 public class Search {
-  public boolean search(int [] array, int target) {
+  public boolean contains(int [] array, int target) {
     boolean found = false;
     int i;
