+++ main-module/src/main/java/com/diffblue/javatest/nestedobjects/subpackage/Item.java
@@ -1,0 +1,5 @@
+package com.diffblue.javatest.nestedobjects.subpackage;
+
+public class Item {
+  public int cost;
+}
