--- main-module/src/main/java/com/diffblue/javatest/somestuff/B.java
@@ -1,6 +1,0 @@
-package com.diffblue.javatest.somestuff;
-
-public class B {
-  public int x;
-  private int y;
-}
