+++ main-module/src/test/java/com/diffblue/javatest/CalcTest.java
@@ -5,8 +5,8 @@ import org.junit.Test;
 
 public class CalcTest {
   @Test
-  public void div0() {
-    int r = new Calc().div(2, 1);
-    Assert.assertEquals(r, 2);
+  public void divideByOne() {
+    int result = new Calc().div(2, 1);
+    Assert.assertEquals(result, 2);
   }
 }
